updated common code for respondent
