+++ src/main/steps/common/confirm-contact-details/checkanswers/template.njk
@@ -1,1 +1,9 @@
 {% extends "common/summary/template.njk" %}
+{% block page_content %}
+{% if section %}
+    <span class="govuk-heading-xl govuk-!-margin-bottom-50px">{{ section }}</span>
+  {% endif %}
+  {% if title %}
+    <h1 class="govuk-hint">{{ title }}</h1>
+  {% endif %}
+{% endblock %}
