+++ src/main/steps/common/summary/utils.test.ts
@@ -1,7 +1,7 @@
 import mockUserCase from '../../../../test/unit/utils/mockUserCase';
 import { MIAM_START } from '../../urls';
 
-import { summaryList } from './utils';
+import { SummaryList, summaryList } from './utils';
 
 const enContent = {
   section: 'Check your details',
@@ -28,8 +28,15 @@ describe('common > summary > utils', () => {
       {
         userCase: mockUserCase,
         expected: {
+          title: 'applicationDetails',
           rows: [
             {
+              key: {
+                text: 'What is a Mediation Information and Assessment Meeting (MIAM)?',
+              },
+              value: {
+                html: 'Yes',
+              },
               actions: {
                 items: [
                   {
@@ -39,17 +46,24 @@ describe('common > summary > utils', () => {
                   },
                 ],
               },
-              key: { text: 'What is a Mediation Information and Assessment Meeting (MIAM)?' },
-              value: { html: 'Yes' },
             },
           ],
-          title: 'applicationDetails',
         },
       },
     ])('return correct summary list items when %#', ({ userCase, expected }) => {
-      expect(summaryList(enContent, userCase, urls, 'applicationDetails', enContent.fieldType, 'en')).toStrictEqual(
-        expected
+      const result: SummaryList | undefined = summaryList(
+        enContent,
+        userCase,
+        urls,
+        'applicationDetails',
+        enContent.fieldType,
+        'en',
+        undefined
       );
+      console.log('======>' + JSON.stringify(result));
+      expect(
+        summaryList(enContent, userCase, urls, 'applicationDetails', enContent.fieldType, 'en', undefined)
+      ).toStrictEqual(expected);
     });
   });
 });
