+++ src/main/steps/respondent/confirm-contact-details/checkanswers/content.ts
@@ -1,13 +1,9 @@
+import { FieldPrefix } from '../../../../app/case/case';
 import { TranslationFn } from '../../../../app/controller/GetController';
 import { FormContent } from '../../../../app/form/Form';
 import { CommonContent } from '../../../common/common.content';
 import { getFormattedDate, summaryList } from '../../../common/summary/utils';
-import {
-  RESPONDENT_ADDRESS_DETAILS,
-  RESPONDENT_ADDRESS_HISTORY,
-  RESPONDENT_CONTACT_DETAILS,
-  RESPONDENT_PERSONAL_DETAILS,
-} from '../../../urls';
+//import { ADDRESS_DETAILS, ADDRESS_HISTORY, CONTACT_DETAILS, PERSONAL_DETAILS } from '../../../urls';
 
 export const enContent = {
   section: 'Check your details',
@@ -17,7 +13,7 @@ export const enContent = {
   },
   keys: {
     applicant1FullName: 'Name',
-    applicant1DateOfBirth: 'Date of birth',
+    applicant1DateOfBirthText: 'Date of birth',
     applicant1PlaceOfBirthText: 'Place of birth',
     address: 'Address',
     addressHistory: 'Address history',
@@ -31,10 +27,12 @@ const en = (content: CommonContent) => {
   const userCase = content.userCase!;
   const dob = userCase.applicant1DateOfBirth;
   getFormattedDate(dob);
+  //const userflag = 'respondent';
+  //const newUrls = getNewUrls(urls, userflag);
   return {
     ...enContent,
     language: content.language,
-    sections: [summaryList(cyContent, userCase, urls, '', fieldType, content.language)],
+    sections: [summaryList(cyContent, userCase, urls, '', fieldType, content.language, FieldPrefix.RESPONDENT)],
   };
 };
 
@@ -46,7 +44,7 @@ const cyContent: typeof enContent = {
   },
   keys: {
     applicant1FullName: 'Name',
-    applicant1DateOfBirth: 'Date of birth',
+    applicant1DateOfBirthText: 'Date of birth',
     applicant1PlaceOfBirthText: 'Place of birth',
     address: 'Address',
     addressHistory: 'Address history',
@@ -57,17 +55,17 @@ const cyContent: typeof enContent = {
 };
 
 const urls = {
-  applicant1FullName: RESPONDENT_PERSONAL_DETAILS,
-  applicant1DateOfBirth: RESPONDENT_PERSONAL_DETAILS,
-  applicant1PlaceOfBirthText: RESPONDENT_PERSONAL_DETAILS,
-  address: RESPONDENT_ADDRESS_DETAILS,
-  addressHistory: RESPONDENT_ADDRESS_HISTORY,
-  applicant1PhoneNumber: RESPONDENT_CONTACT_DETAILS,
-  applicant1EmailAddress: RESPONDENT_CONTACT_DETAILS,
+  applicant1FullName: '_PERSONAL_DETAILS',
+  applicant1DateOfBirthText: '_PERSONAL_DETAILS',
+  applicant1PlaceOfBirthText: '_PERSONAL_DETAILS',
+  address: '_ADDRESS_DETAILS',
+  addressHistory: '_ADDRESS_HISTORY',
+  applicant1PhoneNumber: '_CONTACT_DETAILS',
+  applicant1EmailAddress: '_CONTACT_DETAILS',
 };
 const fieldType = {
   applicant1FullName: 'String',
-  applicant1DateOfBirth: 'Date',
+  applicant1DateOfBirthText: 'Date',
   applicant1PlaceOfBirthText: 'String',
   address: 'String',
   addressHistory: 'String',
@@ -80,7 +78,7 @@ const cy: typeof en = (content: CommonContent) => {
   return {
     ...cyContent,
     language: content.language,
-    sections: [summaryList(cyContent, userCase, urls, '', fieldType, content.language)],
+    sections: [summaryList(cyContent, userCase, urls, '', fieldType, content.language, FieldPrefix.RESPONDENT)],
   };
 };
 
@@ -103,3 +101,10 @@ export const generateContent: TranslationFn = content => {
     form,
   };
 };
+// function getNewUrls(urls: any, userflag: any) {
+//   let key: keyof typeof urls;
+//   for (key in urls) {
+//     let value = urls[key];
+//     urls[key] = userflag+value;
+//   }
+// }
